requirements.txt
annotated-types
anyio
asgiref
certifi
click
colorama
distro
Django
django-cors-headers
django-modeltranslation
djangorestframework
h11
httpcore
httptools
httpx
idna
jiter
joblib
numpy
openai
opencv-python-headless
psycopg
psycopg-binary
psycopg-pool
psycopg2-binary
pydantic
pydantic_core
pydub
PyMuPDF
python-dotenv
PyYAML
scikit-learn
scipy
setuptools
sniffio
sqlparse
threadpoolctl
tqdm
typing-inspection
typing_extensions
tzdata
uvicorn
watchfiles
websockets
wheel